+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 ## 2.3.8 -> 2.3.9:
 * Main improvements:
   * New documentation
+  * Move from jshint to eslint
 
 ## 2.3.7 -> 2.3.8:
 * Main improvements: